+++ beatmap/v2/parse.ts
@@ -1,6 +1,5 @@
-import { IDifficulty } from '../../types/beatmap/v2/difficulty.ts';
 import { Difficulty } from './difficulty.ts';
-import { IInfo } from '../../types/beatmap/v2/info.ts';
+import { IInfoSet } from '../../types/beatmap/v2/info.ts';
 import { Info } from './info.ts';
 import { deepCheck } from '../shared/dataCheck.ts';
 import { DifficultyCheck, InfoCheck } from './dataCheck.ts';
@@ -8,13 +7,15 @@ import { DifficultyRanking } from '../shared/difficulty.ts';
 import logger from '../../logger.ts';
 import { IDataCheckOption } from '../../types/beatmap/shared/dataCheck.ts';
 import { compareVersion } from '../shared/version.ts';
+import { IEvent } from '../../types/beatmap/v2/event.ts';
 
 function tag(name: string): string[] {
    return ['v2', 'parse', name];
 }
 
 export function parseDifficulty(
-   data: Partial<IDifficulty>,
+   // deno-lint-ignore no-explicit-any
+   data: Record<string, any>,
    checkData: IDataCheckOption = { enabled: true, throwError: true },
 ): Difficulty {
    logger.tInfo(tag('difficulty'), 'Parsing beatmap difficulty v2.x.x');
@@ -26,15 +27,23 @@ export function parseDifficulty(
       deepCheck(data, DifficultyCheck, 'difficulty', data._version, checkData.throwError);
    }
 
+   data._customData ||= {};
+   if (data._BPMChanges) data._customData_BPMChanges ||= data._BPMChanges;
+   if (data._bpmChanges) data._customData_bpmChanges ||= data._bpmChanges;
+   if (data._bookmarks) data._customData._bookmarks ||= data._bookmarks;
+   if (data._customEvents) data._customEvents ||= data._customEvents;
+   if (data._time) data._customData._time ||= data._time;
+
    if (compareVersion(data._version, '2.5.0') === 'old') {
-      data._events?.forEach((e) => (e._floatValue = 1));
+      data._events?.forEach((e: IEvent) => (e._floatValue = 1));
    }
 
    return new Difficulty(data);
 }
 
 export function parseInfo(
-   data: Partial<IInfo>,
+   // deno-lint-ignore no-explicit-any
+   data: Record<string, any>,
    checkData: IDataCheckOption = { enabled: true, throwError: true },
 ): Info {
    logger.tInfo(tag('info'), 'Parsing beatmap info v2.x.x');
@@ -47,7 +56,7 @@ export function parseInfo(
       deepCheck(data, InfoCheck, 'info', data._version, checkData.throwError);
    }
 
-   data._difficultyBeatmapSets?.forEach((set) => {
+   data._difficultyBeatmapSets?.forEach((set: IInfoSet) => {
       let num = 0;
       set._difficultyBeatmaps?.forEach((a) => {
          if (typeof a._difficultyRank === 'number') {