+++ beatmap/v3/parse.ts
@@ -1,4 +1,3 @@
-import { IDifficulty } from '../../types/beatmap/v3/difficulty.ts';
 import { Difficulty } from './difficulty.ts';
 import { DifficultyCheck } from './dataCheck.ts';
 import { deepCheck } from '../shared/dataCheck.ts';
@@ -10,7 +9,8 @@ function tag(name: string): string[] {
 }
 
 export function parseDifficulty(
-   data: Partial<IDifficulty>,
+   // deno-lint-ignore no-explicit-any
+   data: Record<string, any>,
    checkData: IDataCheckOption = { enabled: true, throwError: true },
 ): Difficulty {
    logger.tInfo(tag('difficulty'), 'Parsing beatmap difficulty v3.x.x');