+++ beatmap/v2/parse.ts
@@ -8,6 +8,7 @@ import logger from '../../logger.ts';
 import { IDataCheckOption } from '../../types/beatmap/shared/dataCheck.ts';
 import { compareVersion } from '../shared/version.ts';
 import { IEvent } from '../../types/beatmap/v2/event.ts';
+import { shallowCopy } from '../../utils/misc.ts';
 
 function tag(name: string): string[] {
    return ['v2', 'parse', name];
@@ -19,6 +20,7 @@ export function parseDifficulty(
    checkData: IDataCheckOption = { enabled: true, throwError: true },
 ): Difficulty {
    logger.tInfo(tag('difficulty'), 'Parsing beatmap difficulty v2.x.x');
+   data = shallowCopy(data);
    if (!data._version?.startsWith('2')) {
       logger.tWarn(tag('difficulty'), 'Unidentified beatmap version');
       data._version = '2.0.0';
@@ -27,7 +29,7 @@ export function parseDifficulty(
       deepCheck(data, DifficultyCheck, 'difficulty', data._version, checkData.throwError);
    }
 
-   data._customData ||= {};
+   data._customData = shallowCopy(data._customData || {});
    if (data._BPMChanges) data._customData_BPMChanges ||= data._BPMChanges;
    if (data._bpmChanges) data._customData_bpmChanges ||= data._bpmChanges;
    if (data._bookmarks) data._customData._bookmarks ||= data._bookmarks;
@@ -35,7 +37,11 @@ export function parseDifficulty(
    if (data._time) data._customData._time ||= data._time;
 
    if (compareVersion(data._version, '2.5.0') === 'old') {
-      data._events?.forEach((e: IEvent) => (e._floatValue = 1));
+      data._events = data._events?.map((e: IEvent) => {
+         e = shallowCopy(e);
+         e._floatValue = 1;
+         return e;
+      }) || [];
    }
 
    return new Difficulty(data);
@@ -47,50 +53,56 @@ export function parseInfo(
    checkData: IDataCheckOption = { enabled: true, throwError: true },
 ): Info {
    logger.tInfo(tag('info'), 'Parsing beatmap info v2.x.x');
+   data = shallowCopy(data);
    if (!data._version?.startsWith('2')) {
       logger.tWarn(tag('info'), 'Unidentified beatmap version');
+      data._version = '2.0.0';
    }
-   // FIXME: temporary fix from my own mistake, remove when 2.2.0 exist
-   data._version = '2.0.0';
    if (checkData.enabled) {
       deepCheck(data, InfoCheck, 'info', data._version, checkData.throwError);
    }
 
-   data._difficultyBeatmapSets?.forEach((set: IInfoSet) => {
-      let num = 0;
-      set._difficultyBeatmaps?.forEach((a) => {
-         if (typeof a._difficultyRank === 'number') {
-            if (a._difficultyRank - num <= 0) {
-               logger.tWarn(tag('info'), a._difficulty + ' is unordered');
+   data._difficultyBeatmapSets =
+      data._difficultyBeatmapSets?.map(shallowCopy).map((set: IInfoSet) => {
+         set = shallowCopy(set);
+         let num = 0;
+         set._difficultyBeatmaps = set._difficultyBeatmaps?.map((diff) => {
+            diff = shallowCopy(diff);
+            if (typeof diff._difficultyRank === 'number') {
+               if (diff._difficultyRank - num <= 0) {
+                  logger.tWarn(tag('info'), diff._difficulty + ' is unordered');
+               }
+            } else if (typeof diff._difficulty === 'string') {
+               diff._difficultyRank = DifficultyRanking[diff._difficulty];
+               if (!diff._difficultyRank) {
+                  diff._difficulty = 'Easy';
+                  diff._difficultyRank = 1;
+               }
+            } else {
+               diff._difficulty = 'Easy';
+               diff._difficultyRank = 1;
             }
-         } else if (typeof a._difficulty === 'string') {
-            a._difficultyRank = DifficultyRanking[a._difficulty];
-            if (!a._difficultyRank) {
-               a._difficulty = 'Easy';
-               a._difficultyRank = 1;
+            if (DifficultyRanking[diff._difficulty!] !== diff._difficultyRank) {
+               logger.tError(tag('info'), diff._difficulty + ' has invalid rank');
             }
-         } else {
-            a._difficulty = 'Easy';
-            a._difficultyRank = 1;
-         }
-         if (DifficultyRanking[a._difficulty!] !== a._difficultyRank) {
-            logger.tError(tag('info'), a._difficulty + ' has invalid rank');
-         }
-         num = a._difficultyRank;
-         if (
-            typeof a._customData?._editorOffset === 'number' &&
-            a._customData._editorOffset === 0
-         ) {
-            delete a._customData._editorOffset;
-         }
-         if (
-            typeof a._customData?._editorOldOffset === 'number' &&
-            a._customData._editorOldOffset === 0
-         ) {
-            delete a._customData._editorOldOffset;
-         }
-      });
-   });
+            num = diff._difficultyRank;
+            diff._customData = shallowCopy(diff._customData || {});
+            if (
+               typeof diff._customData?._editorOffset === 'number' &&
+               diff._customData._editorOffset === 0
+            ) {
+               delete diff._customData._editorOffset;
+            }
+            if (
+               typeof diff._customData?._editorOldOffset === 'number' &&
+               diff._customData._editorOldOffset === 0
+            ) {
+               delete diff._customData._editorOldOffset;
+            }
+            return diff;
+         });
+         return set;
+      }) || [];
 
    return new Info(data);
 }