+++ types/beatmap/mod.ts
@@ -5,7 +5,7 @@
  * ```ts
  * Beatmap V1: 1.5.0
  * Beatmap V2: 2.6.0
- * Beatmap V3: 3.2.0
+ * Beatmap V3: 3.3.0
  * ```
  *
  * Wrapper is used within the library for cross-version compatibility reason.